nccl-ep: Fix LL/rank-major int32 overflow for ~8K batch sizes
Previously, the expert-major layout was fixed, but
the rank-major was assumed to be much less memory-demanding.
However, at 16K tokens, 8K batch size and 64 ranks, the
offset order reaches `64 x 8K x 16K = 8 GB` which
overflows the int32 capacity.


Mirrored-from: 0a913df800942f829790c1fbe5ef87eca2d878b0

+++ contrib/nccl_ep/device/low_latency.cu
@@ -1758,8 +1758,12 @@ __global__ __launch_bounds__(1024, 1) void combine(// INPUT
                                 : 0;
                     if (j_eff < 0) continue; // no local expert found (shouldn't happen)
 
-                    int sndTokenOffset = slot * numBytesPerSlot;
-                    int rcvTokenOffset = (tokenIdx * numTopk + j_eff) * numBytesPerSlot;
+                    // Byte offsets scale with slot (= srcRank*maxTokensPerRank + i)
+                    // and tokenIdx*numTopk, which times numBytesPerSlot can exceed
+                    // INT_MAX — use size_t to avoid 32-bit truncation (matches the
+                    // expert-major path above).
+                    size_t sndTokenOffset = static_cast<size_t>(slot) * numBytesPerSlot;
+                    size_t rcvTokenOffset = static_cast<size_t>(tokenIdx * numTopk + j_eff) * numBytesPerSlot;
 
                     const auto srcDataInt4Ptr = static_cast<const int4*>(inData) + (int64_t)slot * hiddenBf16Int4;
                     const auto sendBufUint8   = static_cast<uint8_t*>(sendBuf) + sndTokenOffset;
